Make a dedicated fs-breadcrumbs component
