+++ ui/app/controllers/allocations/allocation/task/fs.js
@@ -1,7 +1,6 @@
 import Controller from '@ember/controller';
 import { computed } from '@ember/object';
 import { filterBy } from '@ember/object/computed';
-import { isEmpty } from '@ember/utils';
 
 export default Controller.extend({
   queryParams: {
@@ -42,33 +41,4 @@ export default Controller.extend({
       }
     }
   ),
-
-  breadcrumbs: computed('path', function() {
-    const breadcrumbs = this.path
-      .split('/')
-      .reject(isEmpty)
-      .reduce((breadcrumbs, pathSegment, index) => {
-        let breadcrumbPath;
-
-        if (index > 0) {
-          const lastBreadcrumb = breadcrumbs[index - 1];
-          breadcrumbPath = `${lastBreadcrumb.path}/${pathSegment}`;
-        } else {
-          breadcrumbPath = pathSegment;
-        }
-
-        breadcrumbs.push({
-          name: pathSegment,
-          path: breadcrumbPath,
-        });
-
-        return breadcrumbs;
-      }, []);
-
-    if (breadcrumbs.length) {
-      breadcrumbs[breadcrumbs.length - 1].isLast = true;
-    }
-
-    return breadcrumbs;
-  }),
 });
