+++ ui/app/templates/allocations/allocation/task/fs.hbs
@@ -3,42 +3,12 @@
   {{#if task.isRunning}}
     {{#if isFile}}
       {{#task-file allocation=task.allocation task=task file=path stat=stat class="fs-explorer"}}
-        <nav class="breadcrumb" data-test-fs-breadcrumbs>
-          <ul>
-            <li class={{if breadcrumbs "" "is-active"}}>
-              {{#link-to "allocations.allocation.task.fs-root" task.allocation task activeClass="is-active"}}
-                {{task.name}}
-              {{/link-to}}
-            </li>
-            {{#each breadcrumbs as |breadcrumb|}}
-              <li class={{if breadcrumb.isLast "is-active"}}>
-                {{#link-to "allocations.allocation.task.fs" task.allocation task breadcrumb.path activeClass="is-active"}}
-                  {{breadcrumb.name}}
-                {{/link-to}}
-              </li>
-            {{/each}}
-          </ul>
-        </nav>
+        {{fs-breadcrumbs task=task path=path}}
       {{/task-file}}
     {{else}}
       <div class="fs-explorer boxed-section">
         <div class="boxed-section-head">
-          <nav class="breadcrumb" data-test-fs-breadcrumbs>
-            <ul>
-              <li class={{if breadcrumbs "" "is-active"}}>
-                {{#link-to "allocations.allocation.task.fs-root" task.allocation task activeClass="is-active"}}
-                  {{task.name}}
-                {{/link-to}}
-              </li>
-              {{#each breadcrumbs as |breadcrumb|}}
-                <li class={{if breadcrumb.isLast "is-active"}}>
-                  {{#link-to "allocations.allocation.task.fs" task.allocation task breadcrumb.path activeClass="is-active"}}
-                    {{breadcrumb.name}}
-                  {{/link-to}}
-                </li>
-              {{/each}}
-            </ul>
-          </nav>
+          {{fs-breadcrumbs task=task path=path}}
         </div>
         {{#list-table
           source=sortedDirectoryEntries
